requirements.txt
seaborn
scipy
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
pandas
nltk
contractions
scikit-multilearn
scikit-learn
matplotlib
wordcloud